feat: add ProfilePage component with user details and stats
- Added a new ProfilePage component to display user information, including:
  - User avatar, name, college, and role.
  - User stats such as semester, year, followers, streak, etc.
  - Achievements and uploads sections with placeholders.
- Included utility function to generate random avatar background colors.
- Integrated navigation to edit profile page and back button functionality.
